+++ src/Components/Event/List/EventListPage.tsx
@@ -5,7 +5,7 @@ import RecruitingEvents from "./RecruitingEvents";
 import FinishedEvents from "./FinishedEvents";
 import RadioButtons from "./RadioButtons";
 import { Progress } from "./types";
-import CheckLogin from "../../CheckLogin";
+import { getAccessToken } from "../../../Api/Util/token";
 
 export default function EventListPage() {
   const [selectedTab, setSelectedTab] = useState<Progress>("진행중");
@@ -39,46 +39,52 @@ export default function EventListPage() {
     }
   };
 
+  if (!getAccessToken()) {
+    return (
+      <div className="flex justify-center items-center h-screen">
+        <h2>로그인 후 이용해주세요.</h2>
+      </div>
+    );
+  }
+
   return (
-    <CheckLogin>
-      <div className="p-4 m-8">
-        <Tabs selectedTab={selectedTab} onTabChange={setSelectedTab} />
-        <div className="border-b-2 border-r-2 shadow-md">
-          <div className="p-4">
-            <form
-              className="flex justify-between items-center mx-4 my-4"
-              onSubmit={(e) => e.preventDefault()}
-            >
-              {selectedTab === "종료된 행사" && (
-                <div className="flex justify-between w-full">
-                  <label htmlFor="search" className="sr-only">
-                    특정 기간 내 행사 찾기
-                  </label>
-                  <input
-                    type="text"
-                    id="search"
-                    placeholder="특정 기간 내 행사 찾기"
-                    className="border p-2 rounded w-3/4"
-                  />
-                  <RadioButtons
-                    sortOrder={sortOrder}
-                    onSortOrderChange={setSortOrder}
-                  />
-                </div>
-              )}
-              {selectedTab !== "종료된 행사" && (
-                <div className="flex justify-end w-full">
-                  <RadioButtons
-                    sortOrder={sortOrder}
-                    onSortOrderChange={setSortOrder}
-                  />
-                </div>
-              )}
-            </form>
-            <div className="mx-4">{renderTabContent()}</div>
-          </div>
+    <div className="p-4 m-8">
+      <Tabs selectedTab={selectedTab} onTabChange={setSelectedTab} />
+      <div className="border-b-2 border-r-2 shadow-md">
+        <div className="p-4">
+          <form
+            className="flex justify-between items-center mx-4 my-4"
+            onSubmit={(e) => e.preventDefault()}
+          >
+            {selectedTab === "종료된 행사" && (
+              <div className="flex justify-between w-full">
+                <label htmlFor="search" className="sr-only">
+                  특정 기간 내 행사 찾기
+                </label>
+                <input
+                  type="text"
+                  id="search"
+                  placeholder="특정 기간 내 행사 찾기"
+                  className="border p-2 rounded w-3/4"
+                />
+                <RadioButtons
+                  sortOrder={sortOrder}
+                  onSortOrderChange={setSortOrder}
+                />
+              </div>
+            )}
+            {selectedTab !== "종료된 행사" && (
+              <div className="flex justify-end w-full">
+                <RadioButtons
+                  sortOrder={sortOrder}
+                  onSortOrderChange={setSortOrder}
+                />
+              </div>
+            )}
+          </form>
+          <div className="mx-4">{renderTabContent()}</div>
         </div>
       </div>
-    </CheckLogin>
+    </div>
   );
 }
