--- src/Components/Event/List/Interfaces.ts
@@ -1,16 +1,0 @@
-export interface Event {
-  id: number;
-  name: string;
-  mainImageUrl: string;
-  openDate: string;
-  closeDate: string;
-  recruitStartDate: string;
-  recruitEndDate: string;
-}
-
-export interface EventResponse {
-  hasNext: boolean;
-  sliceNumber: number;
-  numberOfElements: number;
-  content: Event[];
-}
